Move MonitoredEditor framework project to UI namespace
