+++ bundles/extensions/emfdomain/tools.vitruv.extensions.emf.monitorededitor/META-INF/MANIFEST.MF
@@ -12,7 +12,7 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.emf.edit,
  org.eclipse.gmf.runtime.diagram.ui;bundle-version="1.7.0",
  org.eclipse.emf.ecore,
- tools.vitruv.framework.monitorededitor,
+ tools.vitruv.framework.ui.monitorededitor,
  org.eclipse.core.expressions,
  org.eclipse.core.resources,
  org.eclipse.gmf.runtime.diagram.ui.resources.editor;bundle-version="1.7.0",
