+++ bundles/framework/tools.vitruv.framework.monitorededitor/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>tools.vitruv.framework.monitorededitor</name>
+	<name>tools.vitruv.framework.ui.monitorededitor</name>
 	<comment></comment>
 	<projects>
 	</projects>
