--- bundles/framework/tools.vitruv.framework.monitorededitor/src/tools/vitruv/framework/monitorededitor/AbstractMonitoredEditor.java
@@ -1,12 +1,0 @@
-package tools.vitruv.framework.monitorededitor;
-
-import tools.vitruv.framework.vsum.VirtualModel;
-
-public abstract class AbstractMonitoredEditor {
-    protected final VirtualModel virtualModel;
-
-    public AbstractMonitoredEditor(final VirtualModel virtualModel) {
-        this.virtualModel = virtualModel;
-    }
-
-}
