+++ bundles/framework/tools.vitruv.framework.monitorededitor/src/tools/vitruv/framework/ui/monitorededitor/AbstractMonitoredEditor.java
@@ -1,0 +1,12 @@
+package tools.vitruv.framework.ui.monitorededitor;
+
+import tools.vitruv.framework.vsum.VirtualModel;
+
+public abstract class AbstractMonitoredEditor {
+    protected final VirtualModel virtualModel;
+
+    public AbstractMonitoredEditor(final VirtualModel virtualModel) {
+        this.virtualModel = virtualModel;
+    }
+
+}
