+++ bundles/framework/tools.vitruv.framework.monitorededitor/src/tools/vitruv/framework/ui/monitorededitor/registries/MonitoredProjectsRegistry.java
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.monitorededitor.registries;
+package tools.vitruv.framework.ui.monitorededitor.registries;
 
 /**
  * Registry for monitored projects, which implements the singleton pattern.
