+++ bundles/framework/tools.vitruv.framework.monitorededitor/src/tools/vitruv/framework/ui/monitorededitor/registries/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains classes for registries used over all monitored editors.
  */
-package tools.vitruv.framework.monitorededitor.registries;
+package tools.vitruv.framework.ui.monitorededitor.registries;
