+++ bundles/framework/tools.vitruv.framework.ui.vsum/META-INF/MANIFEST.MF
@@ -14,7 +14,7 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.xtend.lib,
  org.eclipse.xtend.lib.macro,
  tools.vitruv.framework.vsum,
- tools.vitruv.framework.monitorededitor,
+ tools.vitruv.framework.ui.monitorededitor,
  tools.vitruv.framework.tests.util,
  org.apache.log4j
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
